mock_project/requirements.txt
django-flexible-content
django-nose
coverage
